Fixed reducer class

+++ src/main/java/MaxFilmRatingReducer.java
@@ -4,11 +4,11 @@ import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class MaxFilmRatingReducer extends Mapper<LongWritable, Text, Text, DoubleWritable> extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class MaxFilmRatingReducer extends Mapper<LongWritable, Text, Text, DoubleWritable> extends Reducer<Text, DoubleWritable, Text, DoubleWritable> {
 
 	@Override
 	public void reduce(Text key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
-		int maxValue = Integer.MIN_VALUE;
+		double maxValue = Integer.MIN_VALUE;
 		for (DoubleWritable value : values) {
 			maxValue = Math.max(maxValue, value.get());
 	    }
